requirements.txt
requests
boto3
kafka-python
python-dotenv
pyspark==3.5.3
